test(runtime): split install-state integration coverage
