+++ vibeguard-runtime/tests/common/mod.rs
@@ -1,10 +1,12 @@
 #![allow(dead_code)]
 
+use serde_json::Value;
+use std::fs;
 use std::io::Write;
 #[cfg(unix)]
 use std::os::unix::fs::PermissionsExt;
-use std::path::PathBuf;
-use std::process::{Command, Stdio};
+use std::path::{Path, PathBuf};
+use std::process::{Command, Output, Stdio};
 
 pub fn bin() -> Command {
     Command::new(env!("CARGO_BIN_EXE_vibeguard-runtime"))
@@ -19,6 +21,46 @@ pub fn unique_temp_dir(label: &str) -> PathBuf {
             .unwrap()
             .as_nanos()
     ))
+}
+
+pub fn run_with_home(root: &Path, args: &[&str]) -> Output {
+    bin()
+        .args(args)
+        .env("HOME", root.join("home"))
+        .current_dir(root)
+        .output()
+        .expect("vibeguard-runtime command should run")
+}
+
+pub fn assert_output(output: &Output, code: i32, stdout: &str, stderr: &str) {
+    assert_eq!(output.status.code(), Some(code));
+    assert_eq!(String::from_utf8_lossy(&output.stdout), stdout);
+    assert_eq!(String::from_utf8_lossy(&output.stderr), stderr);
+}
+
+pub fn write_json(path: &Path, value: &Value) {
+    if let Some(parent) = path.parent() {
+        fs::create_dir_all(parent).expect("JSON parent should be created");
+    }
+    fs::write(
+        path,
+        format!(
+            "{}\n",
+            serde_json::to_string_pretty(value).expect("fixture should serialize")
+        ),
+    )
+    .expect("JSON fixture should be written");
+}
+
+pub fn read_json(path: &Path) -> Value {
+    serde_json::from_slice(&fs::read(path).expect("state file should be readable"))
+        .expect("state file should contain JSON")
+}
+
+pub fn path_text(path: &Path) -> String {
+    path.to_str()
+        .expect("temporary paths should be UTF-8")
+        .to_string()
 }
 
 #[cfg(unix)]
